Rename hot reload module
